+++ packages/zwave-js/src/lib/commandclass/MeterCC.ts
@@ -15,7 +15,14 @@ import type { ValueID } from "@zwave-js/core";
 import { getEnumMemberName, num2hex } from "@zwave-js/shared";
 import type { Driver } from "../driver/Driver";
 import log from "../log";
-import { CCAPI } from "./API";
+import {
+	CCAPI,
+	SetValueImplementation,
+	SET_VALUE,
+	throwUnsupportedProperty,
+	throwUnsupportedPropertyKey,
+	throwWrongValueType,
+} from "./API";
 import {
 	API,
 	CCCommand,
@@ -103,7 +110,14 @@ export function getSupportedRateTypesValueId(endpoint: number): ValueID {
 	};
 }
 
-// @noSetValueAPI This CC is read-only
+export function getResetValueId(endpoint: number, type?: number): ValueID {
+	return {
+		commandClass: CommandClasses.Meter,
+		endpoint,
+		property: "reset",
+		propertyKey: type,
+	};
+}
 
 @API(CommandClasses.Meter)
 export class MeterCCAPI extends CCAPI {
@@ -147,6 +161,41 @@ export class MeterCCAPI extends CCAPI {
 	}
 
 	// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+	public async getAll() {
+		const node = this.endpoint.getNodeUnsafe()!;
+
+		if (this.version >= 2) {
+			const supportedScales =
+				node.getValue<number[]>(
+					getSupportedScalesValueId(this.endpoint.index),
+				) ?? [];
+			const supportedRateTypes =
+				node.getValue<RateType[]>(
+					getSupportedRateTypesValueId(this.endpoint.index),
+				) ?? [];
+
+			const rateTypes = supportedRateTypes.length
+				? supportedRateTypes
+				: [undefined];
+
+			const ret = [];
+			for (const rateType of rateTypes) {
+				for (const scale of supportedScales) {
+					ret.push(
+						await this.get({
+							scale,
+							rateType,
+						}),
+					);
+				}
+			}
+			return ret;
+		} else {
+			return [await this.get()];
+		}
+	}
+
+	// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 	public async getSupported() {
 		this.assertSupportsCommand(MeterCommand, MeterCommand.SupportedGet);
 
@@ -175,6 +224,39 @@ export class MeterCCAPI extends CCAPI {
 		});
 		await this.driver.sendCommand(cc);
 	}
+
+	protected [SET_VALUE]: SetValueImplementation = async (
+		{ property, propertyKey },
+		value,
+	): Promise<void> => {
+		if (property !== "reset") {
+			throwUnsupportedProperty(this.ccId, property);
+		} else if (
+			propertyKey != undefined &&
+			typeof propertyKey !== "number"
+		) {
+			throwUnsupportedPropertyKey(this.ccId, property, propertyKey);
+		} else if (value !== true) {
+			throwWrongValueType(
+				this.ccId,
+				property,
+				"true",
+				value === false ? "false" : typeof value,
+			);
+		}
+
+		const resetOptions: MeterCCResetOptions =
+			propertyKey != undefined
+				? {
+						type: propertyKey,
+						targetValue: 0,
+				  }
+				: {};
+		await this.reset(resetOptions);
+
+		// Refresh values
+		await this.getAll();
+	};
 }
 
 @commandClass(CommandClasses.Meter)
@@ -311,6 +393,8 @@ supports reset:       ${supportsReset}`;
 				ret += "_" + getEnumMemberName(RateType, rateType);
 			}
 			return ret;
+		} else if (property === "reset" && typeof propertyKey === "number") {
+			return getMeterTypeName(propertyKey);
 		}
 		return super.translatePropertyKey(property, propertyKey);
 	}
@@ -613,6 +697,37 @@ export class MeterCCSupportedReport extends MeterCC {
 	@ccValue({ internal: true })
 	public get supportedRateTypes(): readonly RateType[] {
 		return this._supportedRateTypes;
+	}
+
+	public persistValues(): boolean {
+		if (!super.persistValues()) return false;
+		if (!this._supportsReset) return true;
+
+		const valueDb = this.getValueDB();
+		// Create reset values
+		if (this.version < 6) {
+			const resetAllValueId: ValueID = getResetValueId(
+				this.endpointIndex,
+			);
+			if (!valueDb.hasMetadata(resetAllValueId)) {
+				this.getValueDB().setMetadata(resetAllValueId, {
+					...ValueMetadata.WriteOnlyBoolean,
+					label: `Reset accumulated values`,
+				});
+			}
+		} else {
+			const resetSingle: ValueID = getResetValueId(
+				this.endpointIndex,
+				this._type,
+			);
+			if (!valueDb.hasMetadata(resetSingle)) {
+				this.getValueDB().setMetadata(resetSingle, {
+					...ValueMetadata.WriteOnlyBoolean,
+					label: `Reset (${getMeterTypeName(this._type)})`,
+				});
+			}
+		}
+		return true;
 	}
 }